Add naive StubWidget.findById to approximate getElementById.
This is solely to make testing easier ("grab me element X") and not meant to
faithfully model the DOM/hierarchy of the actual app. The hierarchy of widgets
within ui.xml files is particularly naive.

Introduces a StubView base class for StubXxxView classes which holds
the naively-flat list of widgets.