Major view generation, package refactoring.
Previously ui.xml files pointed to widgets that didn't implement the IsXxx
interfaces (e.g. stock GWT widgets) and we had to jump through hops to map the
GWT implementation to a Tessell subclass of it, which did implement the IsXxx
interface.

This also meant the ui.xml file basically had the wrong types, so we had to
use UiField(provided = true) and do our own instantiation of the subclass
so that we could return it from the getters.

This worked, but meant nothing that used UiConstructors or custom parsers
could be supported unless Tessell reimplemented UiBinders instantatian
logic.

Now, instead we funge the ui.xml ui:import from pointing to the com.google
package to an org.tessell packge where the subclassed IsXxx implementations
live, and renamed them from GwtXxx to just Xxx so that the ui.xml references
like gwt:TextBox will still line up.

This gets Tessell out of the business of instantiating widgets, and generally
simpler all around.

An upshot of pointing at the IsXxx impl is that, since it knows about Tessell,
it can use an OtherTypes annotation to point tessell to it's IsXxx and StubXxx
classes, meaning we don't have a huge ugly hardcoded map of that information any
more.