Correctly handle derived properties.
Previously we only implicitly registered dependencies on initial
construction of the property.

However, this fails to handle the case where the derived value
conditionally depends on a number of properties, and specifically some
that may have not been evaluated on the initial constructor due to
they're logic not being called (early returns).