Clean up cookie binding.
Rename Cookie -> IsCookie to follow the IsXxx convention.

Make IsCookie extend TakesValue, so we can reuse the to(TakesValue)
method in PropertyBinder.

While we're at it, make Property extend TakesValue, which is a
breaking change due to the previous Value<P> getValue(), which
should not have been public anyway.

Renamed that to getValueObject and made it protected.

Also tweaked the implementation of to(TakesValue) to support the
use case of initializing a property to a cookie's value, but there
after pushing the property back into the cookie.