Remove Dto interface because it was inherently flawed.
The toModel method was unrealistic, at least in typical GWT projects, as the DTO
will be "shared" code, while the model is client code, so it is likely/expected
that the DTO has no knowledge of the client-side model.

We'll have to use a different abstraction, e.g. the ListProperty.as with an
ElementConverter, as then the client-side code can provide the dto -> model
logic, which is fine as it can see both types.