Add MemberChangedEvent to percolate changes up the model tree.
Models convert property events to member changes, and ListProperties, if they
contain models, forward member changes on to their enclosing models.

This means one can listen to a parent model and know when any of the parent's or
children's properties changed.

Also, added a test for derived values that are based on the contents of list,
which are themselves models. Turns out this already worked, because the implicit
dependency tracking setup both the list and any model property the derived value
accessed as upstream properties.