Static-ify Binder.* methods, make Binder un-instantiable.
I had been meaning to do this for awhile, and finally got around to it.

The unfortunate aspect is that, in a presenter, the "Binder.bind" method
conflicts with the "super.bind" method in the Presenter interface, so a static
import cannot be used (e.g. just "bind(property)..."). Of course it still works
with the qualified "Binder.bind", but that is not as sexy.

Given user code very rarely calls "Presenter.bind()", then perhaps it could
be renamed.