Move back to Binder tracking HandlerRegistrations.
Even if unneeded most of the time, it seems safest to squirrel away
HandlerRegistrations that are setup via binding within the Binder
instance, and just automatically unhook them on Presenter.unbind
without the programmer being none the wiser.

This means Binder is back to having instance state, which means
the Presenter.bind/Binder.bind static import clash doesn't happen
any more, and we don't need AbstractBound to extend Binder to
get the static bind/etc. methods within scope.

tl;dr it's back to the way it was before, `binder.bind(...)`.