Add Binder.bind(BooleanProperty).to(List).has(value).
For Lists, the binding is one-way, only the BooleanProperty drives the changes
in the list (except for caveat below).

For ListPropertys, the binding is two-way, any change to the BooleanProperty or
the ListProperty will update the other.

For either Lists or ListProperties, if the BooleanProperty is unset on the
initial binding, then it will be updated to match the list's current state.