Large Rule refactoring so Property.addRule can call reassess().
Previously the AbstractRule constructor added itself to the parent
Property. However, this meant addRule was called with an only partially
instantiating rule, because the rest of the rule instance's constructor logic
(in the AbstractRule subclass) had not run yet.

Which means Property.addRule could not call reassess() without causing
partially-instantiated rules' validate() to be called, which would lead
to NPEs.

Instead, the safer thing to do is remove the side effect from the AbstractRule
constructor, let Rules fully instantiate themselves, and then manually call
Property.addRule.

It is not as sexy, but has better semantics, and means the property can reassess
itself after every addRule. Which makes sense.