Stop presenters pretending to be GWT IsWidgets.
This broke StubSimplePanel, which was assuming that, in unit tests, any GWT
IsWidget must be a Tessell IsWidget. (Which is a valid assumption because
otherwise it's a regular DOM-coupled Widget.)

To do this right, BasicPresenter would have to implement the Tessell IsWidget
interface, which has quite a few methods in it, and delegate them to its view.

Doing this much delegation to a composite is generally a pita/not worthwhile, so
avoiding it for now.