Change set(null) with a default value to fire change.
This gives the opportunity for properties/widgets that are bound to us to see
that "oh, the null value got ignored", and update themselves accordingly.

Specifically, without this, the user might delete a value from a text box (set
to null), which default value would ignore, but the text box was never told
"btw, the new value is not actually null".