When a property is set to null, auto-select first select box value.
This behavior (auto-select first if null) was already triggered on
the initial binding of the property to the list box, but was not
triggered subsequently if the property was set back to null.

This led to a different user experience (this select box can't
be null...oh, now it can...), so go ahead and auto-select on
property change as well.

If the application really wants the property to be allowed to be
null, they can be null in the list of options.