Expand ResourcesGenerator so it will support font awesome css files.
There are two fixes:

1) Font awesome used "url(foo.svg?v=4.1.0)" and the query string would
mess up the convention that ResourcesGenerator assumes any url(foo)
will have a corresponding AppResource.fooData() available.

So we strip the query string out of the url; I'm not entirely sure
if this is okay, e.g. it might be used for browser hacks, but it
looks like.

2) Font awesome has multiple font files with the same name but
different extensions, e.g. foo.svg, foo.eot, foo.woff.

This threw off the ResourcesGenerator convention, as it assumed
url(foo.svg) would have an AppResources.fooData. But now we need
AppResources.fooSvgData, fooEotData, etc. to avoid collisions.

This last fix is a breaking change, so I'm not yet putting this
on master; ideally we could detect when their are multiple files
with the same name, just different extensions, and only conditionally
add the "foo<Extension>Data" to the method name if necessary.

Fixes #21.