Fix Property.changed() to work when bound to text boxes.
Basically, we do need to treat:

* Property.set(newValue, false);
* Property.setInitialValue(newValue);

Differently.

Even though both change the value, without marking the property
as touched, only the 2nd should reset the initial value, which
is what is used to determined "have we changed".

In particular, when bound against a text box, it uses the
"set(newValue, false)" method, and we really do want to recognize
the user's input as a change, of course.